thinkai-voice-agent/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
cartesia==2.0.17
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.5
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
fastapi==0.127.1
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-speech==2.37.0
google-cloud-texttospeech==2.34.0
google-genai==1.66.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
iterators==0.2.0
jiter==0.13.0
joblib==1.5.3
livekit==1.1.2
livekit-agents==1.4.4
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-anthropic==1.4.4
livekit-plugins-cartesia==1.4.4
livekit-plugins-deepgram==1.4.4
livekit-plugins-elevenlabs==1.4.4
livekit-plugins-google==1.4.4
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.4.4
livekit-protocol==1.1.2
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.3
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.24.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pillow==12.1.1
pipecat-ai==0.0.104
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pyloudnorm==0.1.1
pyparsing==3.3.2
pyreadline3==3.5.4
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
resampy==0.4.3
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.1
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.5
soxr==0.5.0.post1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
types-protobuf==6.32.1.20260221
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0
zipp==3.23.0
supabase
